+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Display user description after role update (#117)
+-   CP workflow graph is now built independently of the task categories (#114)
 
 ## [0.35.0] - 2022-09-26